Support empty device name in TUNSETIFF
In Linux there is a common practice of invoking TUNSETIFF request without supplying IFName.
The Linux behavior is to populate the IFName buffer with a newly generated name, i.e. tun%d or tap%d, based  on the requested device type. The %d is replaced with the first available index starting from zero.

Currently gVisor does not support this feature and uses the empty IFName "as-is", which could be considered a bug, as empty interface names are not allowed in Linux.